requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.13
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0